requirements.txt
fastapi
python-dotenv
uvicorn
psycopg2-binary
pydantic